Revert "Workaround inline class not allowed in external interfaces in JS https://youtrack.jetbrains.com/issue/KT-43224 still need to verify there is no narrowing or truncation when not using unsinged inline classes"
Left the checks for UInt in password hash

This reverts commit ce868fc598612f759c130715bc1caa2b21546739.

+++ multiplatform-crypto-libsodium-bindings/src/jsMain/kotlin/com/ionspin/kotlin/crypto/kdf/Kdf.kt
@@ -26,8 +26,8 @@ actual object Kdf {
         masterKey: UByteArray
     ): UByteArray {
         return getSodium().crypto_kdf_derive_from_key(
-            subkeyLength,
-            subkeyId.toInt(),
+            subkeyLength.toUInt(),
+            subkeyId.toUInt(),
             context,
             masterKey.toUInt8Array()
         ).toUByteArray()
